Refactor and documentation
